Rewrote server.js to use MongoDB
