+++ .idea/workspace.xml
@@ -1,13 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="ChangeListManager">
-    <list default="true" id="6d51bfc6-7107-4f11-ab1e-03eb79ea192b" name="Default Changelist" comment="">
-      <change beforePath="$PROJECT_DIR$/.idea/a2-shortstack.iml" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/.idea/modules.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/modules.xml" afterDir="false" />
+    <list default="true" id="6d51bfc6-7107-4f11-ab1e-03eb79ea192b" name="Default Changelist" comment="Rewrote server.js to use MongoDB">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/package.json" beforeDir="false" afterPath="$PROJECT_DIR$/package.json" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/public/js/scripts.js" beforeDir="false" afterPath="$PROJECT_DIR$/public/js/scripts.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/server.improved.js" beforeDir="false" afterPath="$PROJECT_DIR$/server.improved.js" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/server.improved.js" beforeDir="false" afterPath="$PROJECT_DIR$/server.js" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -39,11 +36,40 @@
       <option name="number" value="Default" />
       <option name="presentableId" value="Default" />
       <updated>1600492195678</updated>
-      <workItem from="1600492197160" duration="1889000" />
+      <workItem from="1600492197160" duration="1975000" />
+      <workItem from="1600649507529" duration="9245000" />
+      <workItem from="1600718483001" duration="11661000" />
     </task>
+    <task id="LOCAL-00001" summary="Made server.improved.js into an Express server">
+      <created>1600494176830</created>
+      <option name="number" value="00001" />
+      <option name="presentableId" value="LOCAL-00001" />
+      <option name="project" value="LOCAL" />
+      <updated>1600494176830</updated>
+    </task>
+    <option name="localTasksCounter" value="2" />
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
     <option name="version" value="3" />
+  </component>
+  <component name="VcsManagerConfiguration">
+    <MESSAGE value="Made server.improved.js into an Express server" />
+    <MESSAGE value="Rewrote server.js to use MongoDB" />
+    <option name="LAST_COMMIT_MESSAGE" value="Rewrote server.js to use MongoDB" />
+  </component>
+  <component name="WindowStateProjectService">
+    <state x="393" y="0" key="CommitChangelistDialog2" timestamp="1600733509052">
+      <screen x="0" y="0" width="1536" height="824" />
+    </state>
+    <state x="393" y="0" key="CommitChangelistDialog2/0.0.1536.824@0.0.1536.824" timestamp="1600733509052" />
+    <state x="361" y="145" key="Vcs.Push.Dialog.v2" timestamp="1600494181098">
+      <screen x="0" y="0" width="1536" height="824" />
+    </state>
+    <state x="361" y="145" key="Vcs.Push.Dialog.v2/0.0.1536.824@0.0.1536.824" timestamp="1600494181098" />
+    <state x="425" y="187" key="run.anything.popup" timestamp="1600732640280">
+      <screen x="0" y="0" width="1536" height="824" />
+    </state>
+    <state x="425" y="187" key="run.anything.popup/0.0.1536.824@0.0.1536.824" timestamp="1600732640280" />
   </component>
 </project>
